requirements.txt
asgiref==3.8.1
Django==5.2.1
Jinja2==3.1.6
MarkupSafe==3.0.2
pillow==11.2.1
PyMySQL==1.1.1
sqlparse==0.5.3
tzdata==2025.2
